moving out the accumulator
